refactor: move provider hooks to src/shared/hooks/
Move useCategories, useProjects, and useTasks hooks from provider files
to dedicated hook files in shared/hooks. Update all import paths across
the codebase.
