+++ src/features/categories/index.ts
@@ -1,3 +1,3 @@
 export type { Category } from "./types/category.types";
 export { categoryService } from "./services/category.service";
-export { useCategories } from "@/app/providers/category-provider";
+export { useCategories } from "@/shared/hooks/use-categories";
