+++ src/features/tasks/components/task-add-dialog.tsx
@@ -29,8 +29,8 @@ import { Separator } from "@/shared/components/ui/separator";
 import { TaskDropdownMenu } from "@/features/tasks/components/task-dropdown-menu";
 import type { CalendarTask, Subtask, TaskPriority, ReminderBefore } from "@/features/tasks/types/task.types";
 import { TaskReminderToggle } from "@/features/reminders/components/task-reminder-toggle";
-import { useProjects } from "@/app/providers/project-provider";
-import { useCategories } from "@/app/providers/category-provider";
+import { useProjects } from "@/shared/hooks/use-projects";
+import { useCategories } from "@/shared/hooks/use-categories";
 import { autoSetPriority, generateDescription, hasAiKey } from "@/shared/services/ai.service";
 import { cn } from "@/shared/lib/utils";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/shared/components/ui/tooltip";
