+++ src/shared/hooks/use-tasks.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { TaskContext } from "@/app/context/task-context";
+
+export function useTasks() {
+  const context = useContext(TaskContext);
+  if (!context) {
+    throw new Error("useTasks must be used within a TaskProvider");
+  }
+  return context;
+}
